Refactor server.js into smaller modules
- Moved helper functions to utils/ directory.
- Moved database initialization and pool to db/ directory.
- Moved WebSocket handling to websocket/ directory.
- Moved API route handlers to routes/ directory.
- Updated server.js to import and use new modules.
- Addressed state management for mode and admin functionalities.
