+++ polycast-backend/routes/translationRoutes.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const router = express.Router();
+const llmService = require('../services/llmService');
+
+// API routes
+router.get('/translate/:language/:text', async (req, res) => {
+    try {
+        const { language, text } = req.params;
+        const translatedText = await llmService.translateText(decodeURIComponent(text), language);
+        res.json({ translation: translatedText });
+    } catch (error) {
+        console.error("Translation API error:", error);
+        res.status(500).json({ error: error.message });
+    }
+});
+
+module.exports = router;
